pyproject.toml
[project]
name = "psfnova"
version = "0.1.0"
description = "psfnova font"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "colorama>=0.4.6",
    "more-itertools>=10.8.0",
    "pillow>=12.0.0",
    "types-pillow>=10.2.0.20240822",
]
